Rename purchases directory to billing-purchases
